cleaning up the analysis suspension script a bit more and the suspension table script.
I am commenting out the spanish code for now because I think we can use a better workflow where we get translations in postgres tables. So don't think it is worth debugging and fixing this spanish code for now. Also don't think we need spanish visuals yet as we think through a more fleshed out workflow for translations / visuals.